% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_data.R
\name{create_mona_ref}
\alias{create_mona_ref}
\title{Mona annotation referenced creation}
\usage{
create_mona_ref(
  mona_dir = NULL,
  seurat = NULL,
  assay = NULL,
  counts = NULL,
  meta = NULL,
  anno = NULL,
  name = NULL,
  species = NULL,
  type = c("RNA", "ATAC"),
  norm = c("SCT", "LogNorm", "TFIDF")
)
}
\arguments{
\item{mona_dir}{A Mona directory}

\item{seurat}{A Seurat object}

\item{assay}{Which assay contains the normalized gene counts, AKA the 'data' slot. Usually 'RNA' or 'SCT' depending on how it was processed}

\item{counts}{A matrix of log-norm counts, cells as rows}

\item{meta}{A table of cell metadata, cells as rows}

\item{anno}{A vector of one or more annotations in the dataset you wish to train on}

\item{name}{Path where you want to save the reference}

\item{species}{Species of the dataset. The following are supported: human, mouse, rat, fruitfly, nematode, zebrafish, frog, pig}
}
\value{
A Mona reference object
}
\description{
Function to create models for label transfer
Supply either a Mona directory, a Seurat object with assay, or a log-norm counts matrix and a table of metadata
}
