% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_data.R
\name{markers_mona}
\alias{markers_mona}
\title{Mona marker calculate}
\usage{
markers_mona(
  object,
  metadata = NULL,
  cluster = NULL,
  cells = NULL,
  downsample = 500,
  recorrect_umi = T
)
}
\arguments{
\item{object}{A Seurat object}

\item{metadata}{Metadata column}

\item{cluster}{Group within the metadata}

\item{cells}{List of cell names}

\item{downsample}{Number of cells per group}

\item{recorrect_umi}{Recorrect for integrated objects}
}
\value{
DE results
}
\description{
Function for calculating markers for a single cluster
}
