% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_data.R
\name{mona_annotate}
\alias{mona_annotate}
\title{Mona annotation}
\usage{
mona_annotate(mona_ref, exp, anno)
}
\arguments{
\item{exp}{The current Mona dataset counts}

\item{ref}{A Mona reference object}
}
\value{
A list of labels
}
\description{
#IN TESTING
Function for transferring cell labels from a reference to the current Mona dataset
}
