% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_data.R
\name{save_mona_dir_custom}
\alias{save_mona_dir_custom}
\title{Mona directory creation}
\usage{
save_mona_dir_custom(
  counts = NULL,
  meta = NULL,
  reduct = NULL,
  dir = NULL,
  name = NULL,
  description = NULL,
  species = "human",
  markers = T
)
}
\arguments{
\item{counts}{A matrix of lognorm counts}

\item{meta}{A matrix of cell annotations}

\item{reduct}{A named list of reductions, each reduction being a matrix of dimensions/coordinates}

\item{dir}{A directory path where the dataset will be stored on your file system}

\item{name}{The actual name of the dataset that will be displayed within Mona}

\item{description}{A brief sentence describing the dataset. Not required, but useful when sharing with others}

\item{species}{The species the dataset originated from. Affects some functionality, so needs to be specified. The following are supported: human, mouse, rat, fruitfly, nematode, zebrafish, frog, pig}

\item{markers}{Whether to pre-calculate markers across all annotations. Recommended if you plan to use markers often, but processing time can be long depending on dataset size/complexity.}
}
\value{
A 'Mona directory' that can be loaded into Mona
}
\description{
Function for converting any single cell object for use within Mona
Note that the counts, meta, and reduct must be organized along rows by cell
If pulling data from a Seurat/Signac object, make sure to transpose counts and only keep the 'cell.embeddings' for reductions
}
