% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_data.R
\name{integrate_mona}
\alias{integrate_mona}
\title{Mona integration}
\usage{
integrate_mona(
  counts_list = NULL,
  meta_list = NULL,
  mode = c("sct", "lognorm"),
  qc = TRUE
)
}
\arguments{
\item{counts_list}{A named list of raw counts in matrix-like format}

\item{meta_list}{An optional list of dataframes with rownames as cells and colnames as cell metadata}

\item{mode}{Whether to use SCT or LogNormalize}

\item{qc}{Whether to automatically remove low-quality cells}
}
\value{
An integrated Seurat object
}
\description{
Function for combining multiple datasets into a single batch-corrected Seurat object
The 'sct' mode and quality control are recommended for most cases
}
