% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_data.R
\name{integrate_mona}
\alias{integrate_mona}
\title{Mona integration}
\usage{
integrate_mona(object_list = NULL)
}
\arguments{
\item{object_list}{A list of objects processed using 'process_mona'}
}
\value{
An integrated Seurat object
}
\description{
Function for integrating multiple Seurat objects into a larger, batch-corrected Seurat object
'process_mona' must be run individually on each dataset before integration
}
