% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_data.R
\name{markers_mona}
\alias{markers_mona}
\title{Mona marker calculation}
\usage{
markers_mona(
  exp = NULL,
  meta = NULL,
  anno = NULL,
  group = NULL,
  cells.1 = NULL,
  cells.2 = NULL
)
}
\arguments{
\item{exp}{a cell by gene expression matrix}

\item{meta}{a cell by annotation matrix}

\item{anno}{a specific column in meta}

\item{group}{a specific group/cluster within anno}

\item{cells.1}{list of cells}

\item{cells.2}{second list of cells for comparison}
}
\value{
DE results
}
\description{
Function for calculating markers for a single cluster
}
