% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_data.R
\name{markers_mona}
\alias{markers_mona}
\title{Mona marker calculation}
\usage{
markers_mona(
  object,
  meta_table = NULL,
  metadata = NULL,
  cluster = NULL,
  cells = NULL,
  downsample = 500
)
}
\arguments{
\item{object}{A Seurat object}

\item{meta_table}{Metadata data frame}

\item{metadata}{Metadata column}

\item{cluster}{Group within the metadata}

\item{cells}{List of cell names}

\item{downsample}{Number of cells per group}
}
\value{
DE results
}
\description{
Function for calculating markers for a single cluster
}
