% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mona.R
\name{mona}
\alias{mona}
\title{Launch Mona app}
\usage{
mona(mona_dir = NULL, data_dir = NULL, load_data = TRUE, save_data = TRUE)
}
\arguments{
\item{mona_dir}{A Mona directory, will automatically open at startup}

\item{data_dir}{A directory of Mona directories. Can be browsed and opened under 'View datasets'.}

\item{load_data}{Can users load their own datasets? If hosted, use FALSE and provide all needed datasets using 'data_dir'.}

\item{save_data}{Can users save datasets/sessions? If hosted, use FALSE to make app 'read only', or TRUE to allow anyone to modify datasets.}
}
\description{
Launch Mona app
}
