% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_data.R
\name{process_mona}
\alias{process_mona}
\title{Mona scRNA data processing}
\usage{
process_mona(counts = NULL, meta = NULL, mode = c("sct", "lognorm"), qc = TRUE)
}
\arguments{
\item{counts}{Raw counts in a matrix-like format, cells as columns and genes as rows}

\item{meta}{An optional dataframe with rownames as cells and colnames as cell metadata}

\item{mode}{Whether to use SCT or LogNormalize}

\item{qc}{Whether to automatically remove low-quality cells}
}
\value{
Seurat object with SCT/LogNormalize processing applied
}
\description{
Function for creating a standard Seurat object for a single dataset
The 'sct' mode and quality control are recommended for most cases
}
