% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_data.R
\name{process_mona}
\alias{process_mona}
\title{Mona data processing}
\usage{
process_mona(seurat = NULL)
}
\arguments{
\item{seurat}{A Seurat object}
}
\value{
Seurat object with SCT processing applied
}
\description{
Function for automatic QC based on
}
