% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_data.R
\name{run_mona_qc}
\alias{run_mona_qc}
\title{Mona quality control}
\usage{
run_mona_qc(seurat = NULL)
}
\arguments{
\item{seurat}{A Seurat object}
}
\value{
Seurat object with low-quality cells filtered out, also adds quality information to metadata
}
\description{
Function for automatic QC based on nFeature/nCount/mitochondrial RNA.
Should be run prior to 'process_mona' as this function removes low-quality cells.
}
