% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_data.R
\name{save_mona_dir}
\alias{save_mona_dir}
\title{Mona directory creation}
\usage{
save_mona_dir(
  seurat = NULL,
  assay = NULL,
  atac_assay = NULL,
  counts = NULL,
  meta = NULL,
  coords = NULL,
  dir = NULL,
  name = NULL,
  description = NULL,
  species = "human",
  scores = T,
  markers = F,
  password = NULL
)
}
\arguments{
\item{seurat}{A Seurat/Signac object, such as those generated by 'process_mona' and 'integrate_mona'.}

\item{assay}{Which assay contains RNA data, usually 'RNA' or 'SCT'. Must include normalized counts in the 'data' slot.}

\item{atac_assay}{Which assay contains ATAC data, usually 'ATAC'. Must include fragment files and annotations, and 'nCount_ATAC' must be in metadata.}

\item{counts}{A matrix of lognorm counts.}

\item{meta}{A matrix of cell annotations.}

\item{coords}{A named list of reductions/image data, each being a matrix of dimensions/coordinates.}

\item{dir}{The file name/path where you want to save the dataset.}

\item{name}{The name of the dataset that will be displayed within Mona.}

\item{description}{A brief sentence describing the dataset. Not required, but useful when sharing with others.}

\item{species}{Species of the dataset. The following are supported: human, mouse, rat, fruitfly, nematode, zebrafish, frog, pig.}

\item{scores}{Whether to calculate ranks for gene set scores. Set false if you don't use this feature, as it adds to processing time.}

\item{markers}{Whether to pre-calculate markers. Set true if you use markers often, but processing time can be long for many annotations.}

\item{password}{A password to open the Mona directory. NOT secure, is stored as plain text in 'mona.qs'. Meant for controlling access to datasets when hosting Mona.}
}
\value{
A 'Mona directory' that can be loaded into Mona
}
\description{
Function for preparing a dataset for use within Mona
If using a Seurat/Signac object, provide 'seurat' and 'assay'. Otherwise, provide 'counts', 'meta', and 'coords'
Note that counts, meta, and coords must be organized along rows by cell
}
