% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_data.R
\name{save_mona_dir}
\alias{save_mona_dir}
\title{Mona directory creation}
\usage{
save_mona_dir(
  seurat = NULL,
  assay = NULL,
  counts = NULL,
  meta = NULL,
  reduct = NULL,
  dir = NULL,
  name = NULL,
  description = NULL,
  species = "human",
  markers = T,
  password = NULL
)
}
\arguments{
\item{seurat}{A Seurat/Signac object, such as those generated by 'process_mona' and 'integrate_mona'}

\item{assay}{Which assay contains the normalized counts, AKA the 'data' slot. Usually 'RNA' or 'SCT' depending on how it was processed}

\item{counts}{A matrix of lognorm counts}

\item{meta}{A matrix of cell annotations}

\item{reduct}{A named list of reductions, each reduction being a matrix of dimensions/coordinates}

\item{dir}{A directory path where the dataset will be stored on your file system}

\item{name}{The actual name of the dataset that will be displayed within Mona}

\item{description}{A brief sentence describing the dataset. Not required, but useful when sharing with others}

\item{species}{The species the dataset originated from. Affects some functionality, so needs to be specified. The following are supported: human, mouse, rat, fruitfly, nematode, zebrafish, frog, pig}

\item{markers}{Whether to pre-calculate markers across all annotations. Recommended if you plan to use markers often, but processing time can be long depending on dataset size/complexity.}

\item{password}{A password required to open the Mona directory. NOT secure, it will be stored as plain text in the 'mona.qs' file. Primarily to control who can access different datasets when hosting Mona.}
}
\value{
A 'Mona directory' that can be loaded into Mona
}
\description{
Function for preparing a dataset for use within Mona
If using a Seurat/Signac object, provide 'seurat' and 'assay'. Otherwise, provide 'counts', 'meta', and 'reduct'
Note that counts, meta, and reduct must be organized along rows by cell
}
