% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_data.R
\name{save_mona_dir}
\alias{save_mona_dir}
\title{Mona directory creation}
\usage{
save_mona_dir(
  seurat = NULL,
  dir = NULL,
  name = NULL,
  description = NULL,
  species = "human",
  markers = T
)
}
\arguments{
\item{seurat}{A Seurat object, recommend using those generated by 'process_mona' and 'integrate_mona'}

\item{dir}{The name of the directory where the dataset will be stored on your file system. Make it memorable to differentiate from other datasets.}

\item{name}{The actual name of the dataset, determines the name that will be displayed within Mona.}

\item{description}{A brief sentence describing the dataset. Not required, but useful for tracking important information or when sharing with others.}

\item{species}{The species the dataset originated from. Affects some functions within Mona, so should be specified. Must use one of the following, or an NCBI taxonomy ID: human, mouse, rat, fruitfly, nematode, zebrafish, thale-cress, frog, pig}
}
\value{
A 'Mona directory' that can be loaded into Mona
}
\description{
Function for preparing Seurat object for use within Mona. Creates a directory containing the actual object as a .qs file, as well as expression data saved using BPCells.
Also calculates markers, removes unnecessary data, stores metadata.
}
