% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_data.R
\name{save_mona_dir}
\alias{save_mona_dir}
\title{Mona directory creation}
\usage{
save_mona_dir(
  seurat = NULL,
  assay = NULL,
  dir = NULL,
  name = NULL,
  description = NULL,
  species = "human",
  markers = T
)
}
\arguments{
\item{seurat}{A Seurat object, such as those generated by 'process_mona' and 'integrate_mona'}

\item{assay}{Which assay contains the normalized gene counts, AKA the 'data' slot. Usually 'RNA' or 'SCT' depending on how it was processed}

\item{dir}{A directory path where the dataset will be stored on your file system}

\item{name}{The actual name of the dataset that will be displayed within Mona}

\item{description}{A brief sentence describing the dataset. Not required, but useful when sharing with others}

\item{species}{The species the dataset originated from. Affects some functionality, so needs to be specified. The following are supported: human, mouse, rat, fruitfly, nematode, zebrafish, frog, pig}

\item{markers}{Whether to pre-calculate markers across all annotations. Recommended if you plan to use markers often, but processing time can be long depending on dataset size/complexity.}
}
\value{
A 'Mona directory' that can be loaded into Mona
}
\description{
Function for preparing a Seurat/Signac object for use within Mona
}
