% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_data.R
\name{transfer_mona_data}
\alias{transfer_mona_data}
\title{Transfer Mona metadata}
\usage{
transfer_mona_data(mona_dir = NULL, seurat = NULL)
}
\arguments{
\item{mona_dir}{A Mona directory}

\item{seurat}{A Seurat object}
}
\value{
Seurat object with updated metadata
}
\description{
Function for taking metadata from a Mona directory and applying it to a standard Seurat object, such as when dataset has been annotated
}
